refactor: 拆分事件处理函数到独立文件
