+++ src/events/handleAlphaTabRenderFinished.ts
@@ -1,0 +1,7 @@
+import type { ITabUIManager } from "../ITabUIManager";
+
+export function handleAlphaTabRenderFinished(ui: ITabUIManager, leaf: any) {
+	ui.hideLoadingOverlay();
+	// new Notice("Tab rendered!");
+	leaf?.updateHeader?.();
+}
