+++ src/events/handleAlphaTabRenderStarted.ts
@@ -1,0 +1,5 @@
+import type { ITabUIManager } from "../ITabUIManager";
+
+export function handleAlphaTabRenderStarted(ui: ITabUIManager) {
+	ui.showLoadingOverlay("Rendering sheet...");
+}
